Plot coeff. of determination in press-lat section
A code to generate plots of coefficient of determination in pressure-latitude cross-sections for heating and moistening tendencies using numpy files of true data and predictions. No data interpolation is conduced. The coefficient of determination is estimated using the existing points in longitude-latitude grids from GCM's saved data. Several comments are provided in the code in order to make it as comprehensive as possible.